Added simple lifecycle/transaction management for JDO PersistenceManager.
Details: The PersistenceFilter is a servlet filter for all URLs. It acts as
"finally" for all servlets. It is a convinient place to close the
PersistenceManager - which commits the transaction. Closing the PersistencManager late
has another nice side-effect - it means the JDO entity objects remail attached
throughout the application / REST resource scope. Which means no more lazy loading
exceptions and ugly logic for attaching / detaching and refreshing things.



git-svn-id: http://socialtrivia.svn.beanstalkapp.com/st/trunk@23 be0e0292-c861-4e28-8362-b93d608658ca